Move decorators methods in zds/member/decorator.py
